Storing the page transition type and thumbnail in the menu
We need to do this to determine what is the appropriate page transition needed. Also I've added an if statement to see if link destination is not the same as the actual page we are on. If it is the same and the menu is open then we close the menu.